:recycle: Move login action to account actions
